add backgroundImage property for action sheet buttons

+++ lib/JFFUI/UITableView/JULoadMoreCellscalculator.m
@@ -120,7 +120,7 @@ static const NSUInteger RIPagingDisabled         = 0;
     NSUInteger restOfTheItems_ = self.totalElementsCount - self.currentCount;
 
     float items_count_for_index_path_ = 1 + index_;
-    NSUInteger pages_expected_ = ceil( items_count_for_index_path_ / self.pageSize );
+    NSUInteger pages_expected_ = ceilf( items_count_for_index_path_ / self.pageSize );
     NSUInteger elements_expected_ = pages_expected_ * self.pageSize;
 
     //check if paging disabled
